requirements.txt
google~=2.0.3
webapp2~=2.5.2
GoogleAppEngineCloudStorageClient
GoogleAppEnginePipeline
GoogleAppEngineMapReduce
Graphy
simplejson
mock
mox
pg8000
werkzeug
requests~=2.23.0
chardet==3.0.2
idna==2.5
urllib3>=1.21.1,<1.24
certifi>=2017.4.17

WebTest~=2.0.35